implement delete project

+++ backend/test/infrastructure/persistence/test_json_project_repository.py
@@ -13,7 +13,8 @@ project = ProjectMother.create()
 @pytest.fixture(autouse=True)
 def setup():
     yield
-    shutil.rmtree(f"./{ROOT_DIR_NAME}")
+    if os.path.exists(f"./{ROOT_DIR_NAME}"):
+        shutil.rmtree(f"./{ROOT_DIR_NAME}")
 
 
 def test_save_should_create_project_id_directory():
@@ -44,3 +45,17 @@ def test_get_all_should_return_all_project_existing_in_root_directory():
 
     assert len(found) == 1
     assert found[0].project_id == project.project_id
+
+
+def test_delete_project_should_delete_existing_project_directory():
+    repository.save(project)
+
+    repository.delete_project(project.project_id)
+
+    assert os.path.isfile(f"./{ROOT_DIR_NAME}/{project.project_id}/project.json") is False
+
+
+def test_delete_project_should_do_nothing_when_project_not_found():
+    repository.delete_project("123")
+
+    assert os.path.isfile(f"./{ROOT_DIR_NAME}/123/project.json") is False
